Remove custom node sync for windows
Just use the zip package. However the directory hierarchy in the zip package
results in a path that exceeds windows' path length limit, so it's not
actually possible to extract the whole package. To work around this, just
extract the node executable; we don't need the rest of it for the waterfall
anyway.